+++ packages/core/src/interface.ts
@@ -5,7 +5,6 @@ import {
   ObjectDefinitionOptions,
   IMethodAspect,
   AspectMetadata,
-  ResolveFilter,
   MidwayFrameworkType
 } from '@midwayjs/decorator';
 import { ILogger, LoggerOptions } from '@midwayjs/logger';
@@ -188,9 +187,8 @@ export interface IApplicationContext extends IObjectFactory {
   baseDir: string;
   parent: IApplicationContext;
   props: IProperties;
-  messageSource: IMessageSource;
   dependencyMap: Map<string, ObjectDependencyTree>;
-  ready(): Promise<void>;
+  ready();
   stop(): Promise<void>;
   registerObject(identifier: ObjectIdentifier, target: any);
 }
@@ -224,7 +222,6 @@ export const HTTP_SERVER_KEY = '_midway_http_server';
 export interface IContainerConfiguration {
   namespace: string;
   packageName: string;
-  newVersion: boolean;
   addLoadDir(dir: string);
   addImports(imports: string[], baseDir?: string);
   addImportObjects(importObjects: Record<string, unknown>);
@@ -247,15 +244,12 @@ export interface IResolverHandler {
   beforeEachCreated(target, constructorArgs: any[], context);
   afterEachCreated(instance, context, definition);
   registerHandler(key: string, fn: HandlerFunction);
+  hasHandler(key: string): boolean;
   getHandler(key: string);
 }
 
 export interface IMidwayContainer extends IApplicationContext {
-  load(opts: {
-    loadDir: string | string[];
-    pattern?: string | string[];
-    ignore?: string | string[];
-  });
+  load(module?: any);
   bind<T>(target: T, options?: ObjectDefinitionOptions): void;
   bind<T>(
     identifier: ObjectIdentifier,
@@ -263,15 +257,17 @@ export interface IMidwayContainer extends IApplicationContext {
     options?: ObjectDefinitionOptions
   ): void;
   bindClass(exports, namespace?: string, filePath?: string);
+  getDebugLogger();
+  setFileDetector(fileDetector: IFileDetector);
   registerDataHandler(handlerType: string, handler: (...args) => any);
   createChild(): IMidwayContainer;
-  resolve<T>(target: T): T;
+  // resolve<T>(target: T): T;
   /**
    * 默认不添加创建的 configuration 到 configurations 数组中
    */
-  createConfiguration(): IContainerConfiguration;
-  containsConfiguration(namespace: string): boolean;
-  addConfiguration(configuration: IContainerConfiguration);
+  // createConfiguration(): IContainerConfiguration;
+  // containsConfiguration(namespace: string): boolean;
+  // addConfiguration(configuration: IContainerConfiguration);
   getConfigService(): IConfigService;
   getEnvironmentService(): IEnvironmentService;
   getInformationService(): IInformationService;
@@ -279,7 +275,7 @@ export interface IMidwayContainer extends IApplicationContext {
   getAspectService(): IAspectService;
   getCurrentEnv(): string;
   getResolverHandler(): IResolverHandler;
-  addDirectoryFilter(filter: ResolveFilter[]);
+  // addDirectoryFilter(filter: ResolveFilter[]);
   /**
    * Set value to app attribute map
    * @param key
@@ -292,6 +288,10 @@ export interface IMidwayContainer extends IApplicationContext {
    * @param key
    */
   getAttr<T>(key: string): T;
+}
+
+export interface IFileDetector {
+  run(container: IApplicationContext);
 }
 
 export interface IConfigService {