+++ packages/core/test/baseFramework.test.ts
@@ -1,26 +1,26 @@
 import {
   APPLICATION_KEY,
-  CONFIGURATION_KEY,
-  LIFECYCLE_IDENTIFIER_PREFIX,
+  // CONFIGURATION_KEY,
+  // LIFECYCLE_IDENTIFIER_PREFIX,
   MidwayFrameworkType,
   Provide,
-  resetModule,
+  // resetModule,
 } from '@midwayjs/decorator';
 import { MidwayContextLogger } from '@midwayjs/logger';
 import * as assert from 'assert';
 import * as path from 'path';
 import {
   clearAllModule,
-  clearContainerCache,
+  // clearContainerCache,
   IMidwayApplication,
   IMidwayBootstrapOptions,
   MidwayRequestContainer,
   LightFramework,
 } from '../src';
 import * as mm from 'mm';
-import { LifeCycleTest, LifeCycleTest1, TestBinding } from './fixtures/lifecycle';
+// import { LifeCycleTest, LifeCycleTest1, TestBinding } from './fixtures/lifecycle';
 import sinon = require('sinon');
-import { getCurrentApplicationContext, getCurrentMainApp, getCurrentMainFramework } from '../src/util/contextUtil';
+import { getCurrentApplicationContext, getCurrentMainApp, getCurrentMainFramework } from '../src';
 
 @Provide()
 class TestModule {
@@ -32,7 +32,7 @@ class TestModule {
 describe('/test/baseFramework.test.ts', () => {
   beforeEach(() => {
     clearAllModule();
-    clearContainerCache();
+    // clearContainerCache();
   });
 
   it.skip('should load js directory and auto disable', async () => {
@@ -510,49 +510,49 @@ describe('/test/baseFramework.test.ts', () => {
     mm.restore();
   });
 
-  it('lifecycle should be ok', async () => {
-    const framework = new LightFramework();
-    await framework.initialize({
-      baseDir: path.join(
-        __dirname,
-        './fixtures/base-app/src'
-      ),
-    });
-
-    const container = framework.getApplicationContext();
-    container.registerDataHandler(APPLICATION_KEY, () => {
-      return { hello: 123 };
-    });
-    const cfg = container.createConfiguration();
-    container.bind(TestBinding);
-    cfg.bindConfigurationClass(LifeCycleTest);
-    cfg.bindConfigurationClass(LifeCycleTest1);
-
-    await framework.loadLifeCycles();
-
-    const aa = await container.getAsync<LifeCycleTest>(LIFECYCLE_IDENTIFIER_PREFIX + 'lifeCycleTest');
-    expect(aa.ts).toEqual('hello');
-    expect(aa.ready).toBeTruthy();
-    // container.registerObject('hellotest111', '12312312');
-    expect(container.get('hellotest111')).toEqual('12312312');
-
-    const aa1 = await container.getAsync<LifeCycleTest1>(LIFECYCLE_IDENTIFIER_PREFIX + 'lifeCycleTest1');
-    expect(aa1.tts).toEqual('hello');
-    expect(aa1.ready).toBeTruthy();
-
-    const callback = sinon.spy();
-    mm(console, 'log', (m) => {
-      callback(m);
-    });
-
-    expect(container.registry.hasDefinition(LIFECYCLE_IDENTIFIER_PREFIX + 'lifeCycleTest')).toBeTruthy();
-    await framework.stop();
-    expect(container.registry.hasDefinition(LIFECYCLE_IDENTIFIER_PREFIX + 'lifeCycleTest')).toBeFalsy();
-    expect(callback.withArgs('on stop').calledOnce).toBeTruthy();
-
-    resetModule(CONFIGURATION_KEY);
-    mm.restore();
-  });
+  // it('lifecycle should be ok', async () => {
+  //   const framework = new LightFramework();
+  //   await framework.initialize({
+  //     baseDir: path.join(
+  //       __dirname,
+  //       './fixtures/base-app/src'
+  //     ),
+  //   });
+  //
+  //   const container = framework.getApplicationContext();
+  //   container.registerDataHandler(APPLICATION_KEY, () => {
+  //     return { hello: 123 };
+  //   });
+  //   const cfg = container.createConfiguration();
+  //   container.bind(TestBinding);
+  //   cfg.bindConfigurationClass(LifeCycleTest);
+  //   cfg.bindConfigurationClass(LifeCycleTest1);
+  //
+  //   await framework.loadLifeCycles();
+  //
+  //   const aa = await container.getAsync<LifeCycleTest>(LIFECYCLE_IDENTIFIER_PREFIX + 'lifeCycleTest');
+  //   expect(aa.ts).toEqual('hello');
+  //   expect(aa.ready).toBeTruthy();
+  //   // container.registerObject('hellotest111', '12312312');
+  //   expect(container.get('hellotest111')).toEqual('12312312');
+  //
+  //   const aa1 = await container.getAsync<LifeCycleTest1>(LIFECYCLE_IDENTIFIER_PREFIX + 'lifeCycleTest1');
+  //   expect(aa1.tts).toEqual('hello');
+  //   expect(aa1.ready).toBeTruthy();
+  //
+  //   const callback = sinon.spy();
+  //   mm(console, 'log', (m) => {
+  //     callback(m);
+  //   });
+  //
+  //   expect(container.registry.hasDefinition(LIFECYCLE_IDENTIFIER_PREFIX + 'lifeCycleTest')).toBeTruthy();
+  //   await framework.stop();
+  //   expect(container.registry.hasDefinition(LIFECYCLE_IDENTIFIER_PREFIX + 'lifeCycleTest')).toBeFalsy();
+  //   expect(callback.withArgs('on stop').calledOnce).toBeTruthy();
+  //
+  //   resetModule(CONFIGURATION_KEY);
+  //   mm.restore();
+  // });
 
   it('should get service in a component write with app', async () => {
     const framework = new LightFramework();