+++ packages/web-koa/src/framework.ts
@@ -114,7 +114,7 @@ export abstract class MidwayKoaBaseFramework<
   }
 
   public async loadMidwayController(): Promise<void> {
-    const collector = new WebRouterCollector(this.getBaseDir());
+    const collector = new WebRouterCollector();
     const routerTable = await collector.getRouterTable();
     const routerList = await collector.getRoutePriorityList();